fashion-ecom-site/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Databricks and data access
databricks-sdk==0.20.0
databricks-sql-connector==3.0.0

# ML and embeddings
numpy==1.26.3
pillow==10.2.0

# HTTP and utilities
httpx==0.26.0
python-multipart==0.0.6
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6
databricks-vectorsearch>=0.22
requests>=2.31.0